chore: remove opkg CLI integration and dead code
- Delete src/opkg.rs (opkg_install + fallback chain) and commands/opkg.rs
- Relocate import_skill from commands/opkg.rs to commands/skills.rs (actively used)
- Remove opkg_install from Tauri handler registration in lib.rs
- Delete dead opkgInstall() from tauri.ts (zero callers)
- Remove apply_augmented_path/augmented_path_env from paths.rs (only consumer was opkg)
- Strip opkg mentions from i18n locale strings (en, zh, ja, vi, pt-BR)
- Remove OpenPackage Notes sections from README.md, README_ZH.md, README_ZH_hk.md
- Update ARCHITECTURE.md: rename OpenPackage Registry to Skill Registry
- Clean up PRD references in openwork-server-ownership-minimal-tauri.md
