+++ apps/app/src/app/lib/tauri.ts
@@ -671,10 +671,6 @@ export async function engineInstall(): Promise<ExecResult> {
   return invoke<ExecResult>("engine_install");
 }
 
-export async function opkgInstall(projectDir: string, pkg: string): Promise<ExecResult> {
-  return invoke<ExecResult>("opkg_install", { projectDir, package: pkg });
-}
-
 export async function importSkill(
   projectDir: string,
   sourceDir: string,
